Separate modules to data and web
https://github.com/raviwu/sfg-pet-clinic/issues/1

Recompose the project structure so that modules contain their
own concerns.

Use pom package on the root pom and modules with jar.
