+++ .gitignore
@@ -23,3 +23,4 @@
 /dist/
 /nbdist/
 /.nb-gradle/
+/.vscode/
