+++ pet-clinic-data/src/main/java/org/lwstudio/sfgpetclinic/model/Owner.java
@@ -1,0 +1,5 @@
+package org.lwstudio.sfgpetclinic.model;
+
+public class Owner extends Person {
+
+}
