+++ pet-clinic-data/src/main/java/org/lwstudio/sfgpetclinic/model/Vet.java
@@ -1,0 +1,5 @@
+package org.lwstudio.sfgpetclinic.model;
+
+public class Vet extends Person {
+
+}
