+++ pet-clinic-web/src/main/java/org/lwstudio/sfgpetclinic/SfgPetClinicApplication.java
@@ -1,0 +1,12 @@
+package org.lwstudio.sfgpetclinic;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SfgPetClinicApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SfgPetClinicApplication.class, args);
+	}
+}
