--- src/main/java/org/lwstudio/sfgpetclinic/model/Owner.java
@@ -1,5 +1,0 @@
-package org.lwstudio.sfgpetclinic.model;
-
-public class Owner extends Person {
-
-}
