--- src/main/java/org/lwstudio/sfgpetclinic/model/Person.java
@@ -1,22 +1,0 @@
-package org.lwstudio.sfgpetclinic.model;
-
-public class Person {
-    private String firstName;
-    private String lastName;
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-}
