--- src/main/java/org/lwstudio/sfgpetclinic/model/Pet.java
@@ -1,33 +1,0 @@
-package org.lwstudio.sfgpetclinic.model;
-
-import java.time.LocalDate;
-
-public class Pet {
-    private PetType petType;
-    private Owner owner;
-    private LocalDate birthDate;
-
-    public PetType getPetType() {
-        return petType;
-    }
-
-    public void setPetType(PetType petType) {
-        this.petType = petType;
-    }
-
-    public Owner getOwner() {
-        return owner;
-    }
-
-    public void setOwner(Owner owner) {
-        this.owner = owner;
-    }
-
-    public LocalDate getBirthDate() {
-        return birthDate;
-    }
-
-    public void setBirthDate(LocalDate birthDate) {
-        this.birthDate = birthDate;
-    }
-}
