--- src/main/java/org/lwstudio/sfgpetclinic/model/PetType.java
@@ -1,13 +1,0 @@
-package org.lwstudio.sfgpetclinic.model;
-
-public class PetType {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
