--- src/main/java/org/lwstudio/sfgpetclinic/model/Vet.java
@@ -1,5 +1,0 @@
-package org.lwstudio.sfgpetclinic.model;
-
-public class Vet extends Person {
-
-}
