--- src/test/java/org/lwstudio/sfgpetclinic/SfgPetClinicApplicationTests.java
@@ -1,16 +1,0 @@
-package org.lwstudio.sfgpetclinic;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SfgPetClinicApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
